Camera2: check always for the highest supported video quality
we do not set the default quality per device on overlays. So in most cases
like it is now it would always fetch the default quality. We want always the
highest available.

So check for it.

As well do not check height/width for the 3 main video qualities. There
is no need and either the camera HAL support it or not. This prevends
especially for weird camera HAL's like exynos4 wrong support lists.

Change-Id: Id00f08cf3baf44f6d51a883153f812bac24f1fa6

Conflicts:
	src/com/android/camera/CameraSettings.java